+++ lib/whois/parsers/base_afilias2.rb
@@ -71,13 +71,12 @@ module Whois
 
       property_supported :registrar do
         node("Sponsoring Registrar") do |value|
-          id, name = decompose_registrar(value) ||
-              Whois.bug!(ParserError, "Unknown registrar format `#{value}'")
+          id, name = decompose_registrar(value)
 
-          Parser::Registrar.new(
+          Parser::Registrar.new({
               id:           id,
-              name:         name
-          )
+              name:         name,
+          })
         end
       end
 
@@ -130,6 +129,8 @@ module Whois
       def decompose_registrar(value)
         if value =~ /(.+?) \((.+?)\)/
           [$2, $1]
+        else
+          [nil, value]
         end
       end